Redesign of the launching and waiting for walkers and accumulators
Added support for "walker and accumulator" per thread, by passing
the walker through to the accumulator directly when done.

We still use a queue for walkers/accumulators that are finished and
waiting for work. We can go further and tie walker/accumulator
together further.

Change the way counting is done to now launch walkers when N
have been launched so that we don'r only ount measurements at
the accumulator phase.